adding initial design file
file originally written by alex with names of functions and one line descriptions but everything in it was changed (comments and function parameters) by mary because none of the other teams api's were taken into account. most functions are fully commented but there are some aspects of the api interaction with other teams that are not crystal clear. also the design aspect is alex's job and i dont know how our internal ranking calculation will be done so I couldnt add that description to that function.