LA and DDS communication
This may need to be changed considering I don't know the exact urls for link analysis and document data store but there are get requests written for each. We will need to discuss what TT is looking for and they do not have any specified endpoints right now. To post indexing we need the n-grams from TT but again, I am not sure how to post to them right now, we may need to make a whole separate json with the layout tt is looking to receive.